% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{bias_correction}
\alias{bias_correction}
\title{Title}
\usage{
bias_correction(
  X,
  sigma,
  Psihat,
  Varhat,
  cutoffs,
  symmetric,
  symmetric_p,
  identificationapproach
)
}
\arguments{
\item{identificationapproach}{}
}
\value{

}
\description{
Title
}
