% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered_covariance_estimate.R
\name{clustered_covariance_estimate}
\alias{clustered_covariance_estimate}
\title{Title}
\usage{
clustered_covariance_estimate(g, cluster_ID)
}
\arguments{
\item{cluster_ID}{}
}
\value{

}
\description{
Title
}
