% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_stats.R
\name{descriptive_stats}
\alias{descriptive_stats}
\title{Title}
\usage{
descriptive_stats(
  X,
  sigma,
  identificationapproach,
  name,
  symmetric,
  cluster_ID
)
}
\arguments{
\item{cluster_ID}{}
}
\value{

}
\description{
Title
}
