% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_gmm_objective.R
\name{meta_gmm_objective}
\alias{meta_gmm_objective}
\title{Title}
\usage{
meta_gmm_objective(betap, cutoffs, symmetric, X, sigma, cluster_ID)
}
\arguments{
\item{cluster_ID}{}
}
\value{

}
\description{
Title
}
