% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correction.R
\name{plot_correction}
\alias{plot_correction}
\title{Title}
\usage{
plot_correction(
  X,
  sigma,
  Psihat,
  Varhat,
  cutoffs,
  symmetric,
  symmetric_p,
  Studynames,
  identificationapproach
)
}
\arguments{
\item{identificationapproach}{}
}
\value{

}
\description{
Title
}
