% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubias_meta.R
\name{pubias_meta}
\alias{pubias_meta}
\title{Title}
\usage{
pubias_meta(
  data,
  studynames,
  symmetric = 1,
  symmetric_p = 1,
  cutoffs = 1.96,
  GMM = FALSE,
  print_plots = FALSE
)
}
\arguments{
\item{print_plots}{}
}
\value{

}
\description{
Title
}
