% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_analytic_llh.R
\name{replication_analytic_llh}
\alias{replication_analytic_llh}
\title{Title}
\usage{
replication_analytic_llh(
  nuhat,
  tauhat,
  betap,
  cutoffs,
  symmetric,
  Z,
  sigmaZ2,
  C
)
}
\arguments{
\item{C}{}
}
\value{

}
\description{
Title
}
