% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_gmm_objective.R
\name{replication_gmm_objective}
\alias{replication_gmm_objective}
\title{Title}
\usage{
replication_gmm_objective(betap, cutoffs, symmetric, Z, sigmaZ2)
}
\arguments{
\item{sigmaZ2}{}
}
\value{

}
\description{
Title
}
