% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replication_moments.R
\name{replication_moments}
\alias{replication_moments}
\title{Title}
\usage{
replication_moments(betap, cutoffs, symmetric, Z, sigmaZ2)
}
\arguments{
\item{sigmaZ2}{}
}
\value{

}
\description{
Title
}
