% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_function_normal_cdf.R
\name{step_function_normal_cdf}
\alias{step_function_normal_cdf}
\title{Title}
\usage{
step_function_normal_cdf(X, theta, sigma, betap, cutoffs, symmetric)
}
\arguments{
\item{symmetric}{}
}
\value{

}
\description{
Title
}
