% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_variance_llh.R
\name{variation_variance_llh}
\alias{variation_variance_llh}
\title{Title}
\usage{
variation_variance_llh(
  lambdabar,
  tauhat,
  betap,
  cutoffs,
  symmetric,
  X,
  sigma,
  C,
  numerical_integration
)
}
\arguments{
\item{numerical_integration}{}
}
\value{

}
\description{
Title
}
