% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubias_meta.R
\name{pubias_meta}
\alias{pubias_meta}
\title{Identification of and correction for publication bias in meta studies}
\usage{
pubias_meta(
  data,
  studynames = NULL,
  symmetric = TRUE,
  sign_lvl = 0.05,
  GMM = FALSE,
  print_plots = FALSE,
  print_dashboard = FALSE
)
}
\arguments{
\item{data}{A \verb{n x 3} matrix where the first column contains the treatment effects, the second column the associated
standard errors of the estimates and in the third column an ID going from 1 to \code{k}, where \code{k} is the number of unique studies.}

\item{studynames}{Optional. A vector of type \code{character} containing all the Studynames of size \code{n} in the same order as the argument \code{data}.}

\item{symmetric}{If set to \code{TRUE}, the publication probability is assumed to be symmetric around zero. If set to \code{FALSE}, asymmetry is allowed.}

\item{sign_lvl}{A value indicating the significance level at which the analysis should be done. Ultimately leads to the threshold (z-score) for the steps of the publication probability.
By default, the significance level is set at 5\%, hence \code{0.05}.}

\item{GMM}{If set to \code{TRUE}, the publication probability will be estimated via GMM. By default, it is set to \code{FALSE} which uses the MLE
method for estimation.}

\item{print_plots}{If set to \code{TRUE}, descriptive plots as well as correction plots are printed into the working directory in .pdf format.}

\item{print_dashboard}{If set to \code{TRUE}, additionally to the .pdf plots, a dashboard with the same charts in dynamic format is produced.
The dashboard is saved in the working directory. Only possible if \code{print_plots} is set to \code{TRUE}.}
}
\value{
Returns a list object called \code{pubias_result} with the elements \code{Results} and \verb{Corrected Estimates}.
If specified, the list also includes the descriptive, as well as the correction plots as \code{ggplot} objects.

\code{Results} contains the publication probability (\code{Psihat}), its variance (\code{Varhat}) and robust standard error (\code{se_robust}).

\verb{Corrected Estimates} contains the original estimates and the median unbiased estimates (\code{original}, \code{adj_estimates}) as well as the corrected 95\% confidence bounds (\code{adj_L}, \code{adj_U}). In addition, the Bonferroni corrected 95\% confidence bounds are reported (\code{adj_LB}, \code{adj_UB}).
}
\description{
\code{pubias_meta()} calculates the publication probability either by a Maximum Likelihood or a GMM approach and
then reports the corrected estimates for the estimates of a meta-analysis.  Check package vignette for further information.
}
