% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubias_replication.R
\name{pubias_replication}
\alias{pubias_replication}
\title{Title}
\usage{
pubias_replication(
  data,
  studynames,
  symmertic = 1,
  cutoffs = 1.96,
  GMM = FALSE
)
}
\arguments{
\item{GMM}{}
}
\value{

}
\description{
Title
}
